add a configUtil file
